Finally Fixed It
