+++ week2/starter/ColorPicker/ContentView.swift
@@ -1,50 +1,48 @@
 import SwiftUI
 
 struct ContentView: View {
-  @State private var alertIsVisible: Bool = false
-  @State private var redColor: Double = 0.0
-  @State private var greenColor: Double = 0.0
-  @State private var blueColor: Double = 0.0
-  @State private var foregroundColor = Color(red: 0, green: 0, blue: 0)
-
+  @State public var backgroundColor = Color.white
+  @State public var redValue = 0.0
+  @State public var greenValue = 0.0
+  @State public var blueValue = 0.0
+  
   var body: some View {
-
+    
     VStack {
       Text("Color Picker")
         .font(.largeTitle)
         .foregroundColor(Color("TextColor"))
-      RectView()
+      RoundedRectangle(cornerRadius: 0)
+          .foregroundColor(backgroundColor)
+                  .overlay(RoundedRectangle(cornerRadius: 0)
+                    .stroke(backgroundColor, lineWidth: 10)
+                                    .opacity(0.5))
+                  .padding()
+
         .padding()
-      }
-      VStack {
-        SliderViewRed()
-      }
-      .padding()
-      VStack {
-        SliderViewGreen()
-      }
-      .padding()
-      VStack {
-        SliderViewBlue()
-      }
-      .padding()
-      ButtonView()
     }
+    VStack {
+      SliderViewRed(sliderValue: $redValue)
     }
-    
+    .padding()
+    VStack {
+      SliderViewGreen(sliderValue: $greenValue)
+    }
+    .padding()
+    VStack {
+      SliderViewBlue(sliderValue: $blueValue)
+    }
+    .padding()
+    ButtonView(backgroundColor: $backgroundColor, redValue: redValue, greenValue: greenValue, blueValue: blueValue)
+  }
+}
+
 
 
 
 struct ContentView_Previews: PreviewProvider {
-  static var previews: some View {
-    ContentView()
-    VStack {
-      RectView()
-      SliderViewRed()
-      SliderViewGreen()
-      SliderViewBlue()
-      ButtonView()
-    }
+ static var previews: some View {
+   ContentView()
     ContentView()
       .preferredColorScheme(.dark)
       .previewDevice("Iphone 14 Pro Max")
